backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-experimental>=0.0.49
langchain-openai>=0.0.5
langgraph>=0.0.10
llama-index>=0.9.40
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
html2text>=2020.1.16
pydantic>=2.6.0
tiktoken>=0.6.0
mcp>=0.1.0